Move _get_http_response_filename to network.download
